+++ packages/cli/CHANGELOG.md
@@ -10,6 +10,7 @@
 #### 🚀 Updates
 
 - Added a `hasher.warnOnMissingInputs` setting to `.moon/workspace.yml`.
+- Added a `shell` task option to `moon.yml` that will wrap system tasks in a shell.
 - Added a `tags` setting to `moon.yml` for project categorization.
 - Added a `--tags` option to the `moon query projects` command.
 - Added 4 new token variables: `$date`, `$time`, `$datetime`, and `$timestamp`.